Move local_pdf_loader to ocr directory for simpler imports
Revert spec-level utils approach - keep utilities local to pipelines
