--- spec/__init__.py
@@ -1,1 +1,0 @@
-"""Spec examples and utilities for DataTrove."""
