--- spec/pdf-processing/__init__.py
@@ -1,1 +1,0 @@
-"""PDF processing examples and utilities."""
