--- spec/pdf-processing/ocr/__init__.py
@@ -1,1 +1,0 @@
-"""OCR pipeline examples."""
