+++ spec/pdf-processing/ocr/local_pdf_loader.py
