--- spec/utils/__init__.py
@@ -1,5 +1,0 @@
-"""Shared utilities for spec examples and pipelines."""
-
-from spec.utils.local_pdf_loader import load_pdf_documents
-
-__all__ = ["load_pdf_documents"]
